Implemented special case k = 1 for k-cover. Changed rval type from short to int. Short really doesn't give us anything.

+++ programming/c++/zeta_transform/src/types.h
@@ -2,7 +2,7 @@
 #define TYPES_H
 
 typedef int int_t;
-typedef short rval_t;
+typedef int rval_t;
 
 template <class T>
 class PointerList {
